Misc fixes

+++ internal/migration/queue_service.go
@@ -111,11 +111,35 @@ func (s queueService) Update(ctx context.Context, entry *QueueEntry) error {
 }
 
 func (s queueService) DeleteByUUID(ctx context.Context, id uuid.UUID) error {
-	return s.repo.DeleteByUUID(ctx, id)
+	return transaction.Do(ctx, func(ctx context.Context) error {
+		entry, err := s.repo.GetByInstanceUUID(ctx, id)
+		if err != nil {
+			return fmt.Errorf("Failed to get queue entry %q: %w", id.String(), err)
+		}
+
+		if entry.IsMigrating() {
+			return fmt.Errorf("Cannot delete queue entry %q: Currently in a migration phase: %w", id, ErrOperationNotPermitted)
+		}
+
+		return s.repo.DeleteByUUID(ctx, id)
+	})
 }
 
 func (s queueService) DeleteAllByBatch(ctx context.Context, batch string) error {
-	return s.repo.DeleteAllByBatch(ctx, batch)
+	return transaction.Do(ctx, func(ctx context.Context) error {
+		entries, err := s.repo.GetAllByBatch(ctx, batch)
+		if err != nil {
+			return fmt.Errorf("Failed to get queue entries for batch %q: %w", batch, err)
+		}
+
+		for _, entry := range entries {
+			if entry.IsMigrating() {
+				return fmt.Errorf("Cannot delete queue entry %q: Currently in a migration phase: %w", entry.InstanceUUID.String(), ErrOperationNotPermitted)
+			}
+		}
+
+		return s.repo.DeleteAllByBatch(ctx, batch)
+	})
 }
 
 // GetNextWindow returns the next valid migration window for the instance in the batch.
